Extract CodeActionFeature from EditorEventManager
